refactor(new-ui): split config page into domain-scoped components and hooks
ui/src/pages/config/page.tsx was 1181 LOC containing swarm-config CRUD,
connection management, modal state, onboarding card, and dialogs in one module.

Split by domain:
- useSwarmConfig hook — state + mutations for swarm-config CRUD
- useConnections hook — state + mutations for connection management
- SwarmConfigSection — data grid + filters, delegates logic to useSwarmConfig
- ConnectionsSection — connection list, delegates logic to useConnections
- ConfigEntryDialog — add/edit config entry form
- ConfigDetailDialog — view/reveal/copy secret value
- ConnectionFormDialog — add/edit connection with health-check
- ConnectionCard — single connection card with test/connect actions
- WelcomeCard — first-run onboarding form

page.tsx reduced to 52 LOC pure composition.
